upgrade to beta.28
